dnsdist: Move rules to dnsdist-rules-factory.hh
